Refactor into multiple crates.
 - Refactor authority signatures for Ristretto into mc-fog-sig-authority
 - Refactor report signatures for Ed into mc-fog-sig-authority
 - Implement mc_fog_sig_authority::Verifier for PublicAddress in mc-account-keys
 - Make mc_fog_sig::Verifier the roll-up trait (formerly scheme)
 - Implement Verifier for PublicAddress in mc_fog_sig
