+++ Cargo.toml
@@ -41,6 +41,8 @@ members = [
     "fog/report/validation",
     "fog/report/validation/test-utils",
     "fog/sig",
+    "fog/sig/authority",
+    "fog/sig/report",
     "fog/types",
     "ledger/db",
     "ledger/distribution",
