+++ account-keys/Cargo.toml
@@ -17,6 +17,7 @@ mc-crypto-digestible = { path = "../crypto/digestible", features = ["dalek", "de
 mc-crypto-hashes = { path = "../crypto/hashes" }
 mc-crypto-sig = { path = "../crypto/sig" }
 mc-crypto-keys = { path = "../crypto/keys", default-features = false }
+mc-fog-sig-authority = { path = "../fog/sig/authority" }
 mc-util-from-random = { path = "../util/from-random" }
 mc-util-repr-bytes = { path = "../util/repr-bytes", default-features = false, features = ["alloc"] }
 mc-util-serial = { path = "../util/serial" }
