+++ account-keys/src/domain_separators.rs
@@ -1,7 +1,4 @@
 // Copyright (c) 2018-2021 The MobileCoin Foundation
 
-/// The tag for the signature context
-pub const FOG_AUTHORITY_SIGNATURE_TAG: &[u8; 23] = b"Fog authority signature";
-
 /// Domain separator for hashing a private view key and index into a subaddress.
 pub const SUBADDRESS_DOMAIN_TAG: &str = "mc_subaddress";
