+++ fog/sig/Cargo.toml
@@ -3,17 +3,17 @@ name = "mc-fog-sig"
 version = "1.0.0"
 authors = ["MobileCoin"]
 edition = "2018"
-description = "Create and verify fog authority and report signatures"
+description = "Verify Fog Signatures"
 
 [dependencies]
-mc-account-keys = { path = "../../account-keys", default-features = false }
-mc-attest-core = { path = "../../attest/core", default-features = false }
-mc-crypto-digestible-signature = { path = "../../crypto/digestible/signature", default-features = false }
-mc-crypto-keys = { path = "../../crypto/keys", default-features = false }
+mc-account-keys = { path = "../../account-keys"}
+mc-crypto-keys = { path = "../../crypto/keys" }
 mc-crypto-x509-utils = { path = "../../crypto/x509/utils" }
-mc-fog-types = { path = "../types", default-features = false }
+mc-fog-sig-report = { path = "./report" }
+mc-fog-sig-authority = { path = "./authority" }
+mc-fog-types = { path = "../types" }
 
-displaydoc = { version = "0.1.7", default-features = false }
+displaydoc = { version = "0.1.7" }
 pem = "0.8"
 signature = { version = "1.2.2" }
 x509-signature = "0.5"
