+++ fog/sig/authority/README.md
@@ -1,0 +1,3 @@
+This crate provides the basic trait for creating and verifying fog authority signatures, and implementations for the RistrettoPublic / RistrettoPrivate types.
+
+This crate is designed to work inside an enclave (no-std, no FFI dependencies).
