+++ fog/sig/report/Cargo.toml
@@ -1,0 +1,21 @@
+[package]
+name = "mc-fog-sig-report"
+version = "1.0.0"
+authors = ["MobileCoin"]
+edition = "2018"
+description = "Create and verify fog report signatures"
+
+[dependencies]
+mc-attest-core = { path = "../../../attest/core", default-features = false }
+mc-crypto-digestible-signature = { path = "../../../crypto/digestible/signature", default-features = false }
+mc-crypto-keys = { path = "../../../crypto/keys", default-features = false }
+mc-fog-types = { path = "../../types", default-features = false }
+
+displaydoc = { version = "0.1.7", default-features = false }
+signature = { version = "1.2.2" }
+
+[dev-dependencies]
+mc-util-from-random = { path = "../../../util/from-random" }
+
+rand_core = "0.5"
+rand_hc = "0.2"
