+++ fog/sig/report/README.md
@@ -1,0 +1,3 @@
+This crate provides the basic trait for creating and verifying fog report signatures, and implementations for the Ed25519Public / Ed25519Pair types.
+
+This crate is designed to work inside an enclave (no-std, no FFI dependencies).
